+++ CHANGELOG.md
@@ -4,7 +4,8 @@ Yii Framework 2 debug extension Change Log
 2.1.6 under development
 -----------------------
 
-- Enh #392: `tracePathMappings` property for module (Sarke)
+- Enh #392: Add `tracePathMappings` property to module (Sarke)
+- Enh #379: Better error message when no debug data is found (Sarke)
 
 
 2.1.5 June 04, 2019